requirements-train.txt
torch>=2.1.0
torchvision>=0.16.0
torchmetrics>=1.3.0
scikit-learn>=1.2.0
pandas>=1.5.0
numpy>=1.24.0
albumentations>=1.3.0
opencv-python-headless>=4.8.0
pillow>=9.5.0
tqdm>=4.66.0
torchcam>=0.4.0